\name{funcARRAYTRANS_J}
\alias{funcARRAYTRANS_J}
\title{Creates and transposes an array (JHM)}
\description{
For JHM.
For use with Colonyzer output.
Creates and transposes an array.
}
\usage{
funcARRAYTRANS_J(data_vec,dim)
}
\arguments{
  \item{data_vec}{Growth or Time}
  \item{dim}{Dimensions of array}
}
\value{
Array
}
\keyword{qfaBayes}

