\name{funcMODELHierarchical}
\alias{funcMODELHierarchical}
\title{Saves hierarchical model to disk space for rjags to load}
\description{
For use with rjags
For use with Colonyzer output.
Saves hierarchical model to disk space for rjags to load
}
\usage{
funcMODELHierarchical()
}

\value{
"model1.bug" file is saved
}
\keyword{qfaBayes}

