\name{funcMODELInteraction}
\alias{funcMODELInteraction}
\title{Saves the Interaction hierarchical model to disk space for rjags to load}
\description{
For use with rjags
For use with Colonyzer output.
Saves the Interaction hierarchical model to disk space for rjags to load
}
\usage{
funcMODELInteraction()
}
\arguments{
}
\value{
"model1.bug" file is saved
}
\keyword{qfaBayes}

