\name{funcMODELJoint}
\alias{funcMODELJoint}
\title{Saves hierarchical model to disk space for rjags to load}
\description{
For JHM.
For use with rjags
For use with Colonyzer output.
Saves the Joint hierarchical model to disk space for rjags to load
}
\usage{
funcMODELJoint()
}
\arguments{
}
\value{
"model1.bug" file is saved
}
\keyword{qfaBayes}

