\name{funcNoORF}
\alias{funcNoORF}
\title{Gives number of repeats for each ORF}
\description{
For use with Colonyzer output.
Gives number of repeats for each ORF
}
\usage{
funcNoORF(x,data)
}
\arguments{
  \item{x}{ORF name}
  \item{data}{Colonyzer output}
}
\value{
Number of repeats
}
\keyword{qfaBayes}

