\name{funcNoSum}
\alias{funcNoSum}
\title{Gives running total for the number of repeats for each ORF}
\description{
For use with Colonyzer output.
Gives running total for the number of repeats for each ORF
}
\usage{
funcNoSum(x,NoORF_vec)
}
\arguments{
  \item{x}{number of ORFs to count to}
  \item{NoORF_vec}{NoORF}
}
\value{
Total of number of repeats
}
\keyword{qfaBayes}

