\name{funcNoTime}
\alias{funcNoTime}
\title{Gives number of time points for a repeat}
\description{
For use with Colonyzer output.
Gives number of time points for each repeat
}
\usage{
funcNoTime(x,data)
}
\arguments{
  \item{x}{ID code}
  \item{data}{Colonyzer output}
}
\value{
Number of time points for repeat
}
\keyword{qfaBayes}

