\name{funcPRIORS_J}
\alias{funcPRIORS_J}
\title{Load priors from package or .Priors file}
\description{
For JHM.
For use with Colonyzer output.
Load priors from package or .Priors file.
}
\usage{
funcPRIORS_J(CustomModel,Priors)
}
\arguments{
  \item{CustomModel}{Name of Custom Model in order to locate "CustomModel.Priors"}
  \item{Priors}{Variable to fill with prior values}
}
\value{
Prior values
}
\keyword{qfaBayes}

