\name{funcRowRep}
\alias{funcRowRep}
\title{Adds NA values at the end of a repeat to give consistent row length for an array}
\description{
For use with Colonyzer output.
Adds NA values at the end of a repeat to give consistent row length for an array
}
\usage{
funcRowRep(x,NoTime_vec,data_vec,dimr,dimc)
}
\arguments{
  \item{x}{Index}
  \item{NoTime_vec}{NoTime}
  \item{data_vec}{y or x}
  \item{dimr}{Number of rows in array}
  \item{dimc}{Number of columns in array}
}
\value{
y or x with a consistent row length for an array
}
\keyword{qfaBayes}

