\name{funcSCALING}
\alias{funcSCALING}
\title{Scales data by maximum theroetical value for IOD}
\description{
For use with Colonyzer output.
Scales data by maximum theroetical value for IOD
}
\usage{
funcSCALING(data,vec)
}
\arguments{
  \item{data}{original data}
  \item{vec}{Growth}
}
\value{
Scaled growth
}
\keyword{qfaBayes}

