\name{funcREMOVE}
\alias{funcREMOVE}
\title{Filter by Screen Name, Temprature and Master Plate Number}
\description{
For use with Colonyzer output.
Filter by Screen Name, Temprature and Master Plate Number
}
\usage{
funcREMOVE(data,Screen,Treat,MPlate)
}
\arguments{
  \item{data}{Colonyzer output}
  \item{Screen}{Screen name}
  \item{Treat}{Treatment}
  \item{MPlate}{Master Plate}
}
\value{
Filtered Colonyzer output
}
\keyword{qfaBayes}

\name{funcIDORDER}
\alias{funcIDORDER}
\title{Order by ORF and ID coloumn}
\description{
For use with Colonyzer output.
Order by ORF and ID coloumn
}
\usage{
funcIDORDER(data)
}
\arguments{
  \item{data}{Colonyzer output}
}
\value{
Ordered Colonyzer output
}
\keyword{qfaBayes}
