\name{funcXY}
\alias{funcXY}
\title{Sorts data into array with correct dimensions}
\description{
For use with Colonyzer output.
Sorts data into array with correct dimensions
}
\usage{
funcXY(data,M,N,NoTime_vec,NoSum_vec,dimr,dimc)
}
\arguments{
  \item{data}{Growth or Time}
  \item{M}{M max length of Growth or Time}
  \item{N}{Number of ORFs}
  \item{NoTime_vec}{NoTime}
  \item{NoSum_vec}{NoSum}
  \item{dimr}{Number of rows in array}
  \item{dimc}{Number of columns in array}
}
\value{
Array
}
\keyword{qfaBayes}

