\name{IHM_MDRxMDP_postpro}
\alias{IHM_MDRxMDP_postpro}
\title{Performs post processing for IHM}
\description{
For use with SHM output x2.
Performs post-processing for IHM.
}
\usage{
IHM_MDRxMDP_postpro(SHM_a,SHM_output_a,SHM_b,SHM_output_b)
}
\arguments{
  \item{SHM_a}{SHM control post-processing dataset details}
  \item{SHM_output_a}{SHM control output}
  \item{SHM_b}{SHM control post-processing dataset details}
  \item{SHM_output_b}{SHM query output}
}
\value{
Post-processed SHM output for the IHM.
}
\keyword{qfaBayes}
