\name{plot_IHM_simple}
\alias{plot_IHM_simple}
\title{plot IHM_simple_C output}
\description{
For use with IHM output.
plot IHM_simple_C output
}
\usage{
plot_IHM_simple(IHM_output,SHM)
}
\arguments{
  \item{IHM_output}{IHM output}
  \item{SHM}{post processed data for SHM_a or SHM_b}
}
\value{
IHM plots
}
\keyword{qfaBayes}
