\name{plot_JHM_simple}
\alias{plot_JHM_simplen}
\title{plot JHM_simple_C output}
\description{
For use with JHM output.
plot JHM_simple_C output
}
\usage{
plot_JHM_simple(JHM_output,JHM)
}
\arguments{
  \item{JHM_output}{JHM output}
  \item{JHM}{post processed data for JHM}
}
\value{
JHM plots
}
\keyword{qfaBayes}
