\name{plot_SHM_simple}
\alias{plot_SHM_simple}
\title{plot SHM_simple_C output}
\description{
For use with SHM output.
plot SHM_simple_C output
}
\usage{
plot_SHM_simple(SHM_output,SHM)
}
\arguments{
  \item{SHM_output}{SHM output}
  \item{SHM}{post processed data for SHM}
}
\value{
SHM plots
}
\keyword{qfaBayes}
