\name{qfa.variables}
\alias{qfa.variables}
\title{Gives experiment variables from Colonyzer output}
\description{
For use with Colonyzer output.
Gives experiment variables from Colonyzer output
}
\usage{
qfa.variables(data)
}
\arguments{
  \item{data}{Colonyzer output}
}
\value{
Experiment variables
}
\keyword{qfaBayes}
