\name{qfaplots.H}
\alias{qfaplots.H}
\title{Hierachical logistic curve model plots}
\description{
Creates and stores plots for qfa.Hierarchical() output to .pdf
}
\usage{
qfaplots.H(work,QFA,LinearGaussian=FALSE)
}
\arguments{
  \item{QFA}{Output from qfa.Hierarchical()}
  \item{work}{A unique name to identify the functions output.}
\item{LinearGaussian}{Have you used a linear gaussian model? TRUE or FALSE.}
}
\value{
Creates plots and stores then to .pdf in your working directory.
}
\examples{
# example
# demo(Hierarchical7)
#

}
\keyword{qfaBayes}