\name{qfaplots.I}
\alias{qfaplots.I}
\title{Interaction model plots}
\description{
Creates and stores plots for qfa.Interaction() output to .pdf
}
\usage{
qfaplots.I(QFA,work,defa,defb,alpha_a,alpha_b,gam_b,mu_a,mu_b,tau_a,tau_b)
}
\arguments{
  \item{QFA}{Output from qfa.Interaction(). }
  \item{work}{A unique name to identify the functions output.}
\item{LinearGaussian}{Have you used a linear gaussian model? TRUE or FALSE.}
}
\value{
Creates plots and stores then to .pdf in your working directory.
}
\examples{
# example
# demo(Interaction)
#

}
\keyword{qfaBayes}
