\name{visual_convergence_simple_check_IHM}
\alias{visual_convergence_simple_check_IHM}
\title{Gives a trace plot for one of every parameter type to check convergence of IHM output}
\description{
For use with IHM output.
Gives a trace plot for one of every parameter type to check convergence of IHM output.
}
\usage{
visual_convergence_simple_check_IHM(IHM_output)
}
\arguments{
  \item{IHM_output}{posterior samples output from the IHM MCMC sampler}
}
\value{
Gives trace plots for one of every parameter type
}
\keyword{qfaBayes}

