\name{visual_convergence_simple_check_JHM}
\alias{visual_convergence_simple_check_JHM}
\title{Gives a trace plot for one of every parameter type to check convergence of JHM output}
\description{
For use with JHM output.
Gives a trace plot for one of every parameter type to check convergence of JHM output.
}
\usage{
visual_convergence_simple_check_JHM(JHM_output)
}
\arguments{
  \item{JHM_output}{posterior samples output from the JHM MCMC sampler}
}
\value{
Gives trace plots for one of every parameter type
}
\keyword{qfaBayes}

