\name{visual_convergence_simple_check_SHM}
\alias{visual_convergence_simple_check_SHM}
\title{Gives a trace plot for one of every parameter type to check convergence of SHM output}
\description{
For use with SHM output.
Gives a trace plot for one of every parameter type to check convergence of SHM output.
}
\usage{
visual_convergence_simple_check_SHM(SHM_output)
}
\arguments{
  \item{SHM_output}{posterior samples output from the SHM MCMC sampler}
}
\value{
Gives trace plots for one of every parameter type
}
\keyword{qfaBayes}

