\name{funcARRAYTRANS}
\alias{funcARRAYTRANS}
\title{Creates and transposes an array}
\description{
For use with Colonyzer output.
Creates and transposes an array
}
\usage{
funcARRAYTRANS(data_vec,dim)
}
\arguments{
  \item{data_vec}{y or x}
  \item{dim}{Dimensions of array}
}
\value{
An array
}
\keyword{qfaBayes}

