\name{funcARRAYTRANS_J}
\alias{funcARRAYTRANS_J}
\title{Creates and transposes an Array (JHM)}
\description{
For JHM.
For use with Colonyzer output.
Creates and transposes an Array.
}
\usage{
funcARRAYTRANS_J(data_vec,dim)
}
\arguments{
  \item{data}{Growth or Time}
  \item{dimc}{Dimensions of array}
}
\value{
Array
}
\keyword{qfaBayes}

