\name{funcColORF}
\alias{funcColORF}
\title{Adds NA values at the end of an ORF to give consistent Col length for an array}
\description{
For use with Colonyzer output.
Adds NA values at the end of an ORF to give consistent Col length for an array
}
\usage{
funcColORF(x,NoTime_vec,data_vec,dimr,dimc)
}
\arguments{
  \item{x}{Index}
\item{NoTime_vec}{NoTime}
  \item{data_vec}{y or x}
  \item{dimr}{Number of rows in array}
  \item{dimc}{Number of columns in array}
}
\value{
y or x with consistent Col length for an array
}
\keyword{qfaBayes}

