\name{qfa.variables}
\alias{qfa.variables}
\title{Gives experimental variables from QFA data (colonyzer.read output from the qfa package)}
\description{
For use with Colonyzer output.
Gives experimental variables from QFA data (colonyzer.read output from the qfa package).
}
\usage{
qfa.variables(data)
}
\arguments{
  \item{data}{Colonyzer output}
}
\value{
Experiment variables
}
\keyword{qfaBayes}
