\name{qfaplots.I}
\alias{qfaplots.I}
\title{Interaction model plots}
\description{
Creates and stores plots for qfa.Interaction() output to .pdf
}
\usage{
qfaplots.H(work,QFA,LinearGaussian=FALSE)
\arguments{
  \item{work}{A unique name to identify the functions output.}
  \item{QFA}{Output from qfa.Interaction(). )
\item{LinearGaussian}{Have you used a linear gaussian model? TRUE or FALSE.}
}
\value{
Creates plots and stores then to .pdf in your working directory.
}
\examples{
# example
# demo(Interaction)
#

}
\keyword{qfaBayes}