\name{qfaplots.J}
\alias{qfaplots.J}
\title{joint hierachical logistic curve model plots}
\description{
Creates and stores plots for qfa.Joint() output to .pdf
}
\usage{
qfaplots.H(work,QFA,LinearGaussian=FALSE)
\arguments{
  \item{work}{A unique name to identify the functions output.}
  \item{QFA}{Output from qfa.Joint(). )
\item{LinearGaussian}{Have you used a linear gaussian model? TRUE or FALSE.}
}
\value{
Creates plots and stores then to .pdf in your working directory.
}
\examples{
# example
# demo(Joint)
#

}
\keyword{qfaBayes}