\name{SHM_postpro}
\alias{SHM_postpro}
\title{Proforms post processing for SHM}
\description{
For use with Colonyzer output.
Proforms post processing for SHM.
}
\usage{
SHM_postpro(a,Treatment,Screen,MPlate,remove_row,remove_col)
}
\arguments{
\item{a}{Colonyzer output}
  \item{Treatment}{List of treatments}
  \item{Screen}{List of Screens}
  \item{MPlate}{List of Master Plates}
  \item{remove_row}{plate rows to remove}
  \item{remove_col}{plate columns to remove}
}
\value{
Post processed Colonyzer output for SHM.
}
\keyword{qfaBayes}
