\name{SHM_postpro}
\alias{SHM_postpro}
\title{Gives experiment variables from Colonyzer output}
\description{
For use with Colonyzer output.
Gives experiment variables from Colonyzer output
}
\usage{
SHM_postpro(a,Treat,Screen,MPlate)
}
\arguments{
\item{a}{Colonyzer output}
  \item{Treat}{List of treatments}
  \item{Screen}{List of Screens}
  \item{MPlate}{List of Master Plates}
}
\value{
Post processed Colonyzer output for SHM.
}
\keyword{qfaBayes}
